pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game_scenario_agent"
version = "0.1.0"
description = "A multi-agent game scenario using LangChain and LangGraph"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "httpx>=0.25.0",
    "pydantic>=2.4.2",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "langchain>=0.3.25",
    "vllm>=0.9.1",
    "accelerate>=1.7.0",
    "langchain-huggingface>=0.3.0",
    "langchain-community>=0.3.25",
    "langchain-tavily>=0.2.3",
    "langgraph>=0.4.8",
    "langchain-google-genai>=2.1.5",
    "groq>=0.28.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "tavily-python>=0.7.8",
    "langchain-groq>=0.3.2",
    "langgraph-supervisor>=0.0.27",
    "langgraph-swarm>=0.0.11",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 120
target-version = ["py311"]
